AlimentNutritionalTable refactoring
New Features:
- Added a FoodUtil capable of calculating the total nutrients of a given meal, returning an object of type Aliment.
	> It is also possible to pass a SubFood to be used instead.
- Now AlimentNutritionalTable is able to add a meal total column, using the new FoodUtil that calculates the total nutrients of a given Meal and returns an object of type Aliment to facilitate access to properties. When this Aliment object is indicated with all the macronutrients, an extra column will be added indicating the percentage that that Food represents in that Meal.

+++ questNutriPI/src/model/entities/SubFood.java
@@ -31,7 +31,7 @@ public class SubFood extends SubFoodDAO implements ICopy<SubFood>{
     public Aliment aliment;
 
     @Column(name = "subFood_quantity")
-    public Float quantity;
+    public Double quantity;
 
     @Column(name = "subFood_unityQt")
     public String unityQt;
@@ -55,7 +55,7 @@ public class SubFood extends SubFoodDAO implements ICopy<SubFood>{
      * @param unityQt Recebe a Unidade da Comida Substituta
      * @param obs Recebe as Observações da Comida Substituta
      */
-    public SubFood(Integer idSubFood, Food food, Aliment aliment, Float quantity, String unityQt, String obs) {
+    public SubFood(Integer idSubFood, Food food, Aliment aliment, Double quantity, String unityQt, String obs) {
         super();
     	this.idSubFood = idSubFood;
         this.food = food;
